Lots of changes. :) Would lead to far to mention all. Just to name some...... FIX: click outside urlpopupview closes it FIX: MouseOverButton ( press->out->in->release ) FIX: CONTROL+TAB for switching to other windows ADD: completed statusview with bars and icons ADD: Preferences ADD: OPTION+T -> New Tab ADD: OPTION+N -> New Window ADD: OPTION+W -> Closes Window ( as usual ) ADD: OPTION+Q -> Closes Themis
git-svn-id: svn://svn.code.sf.net/p/themis/code@214 ecaf47ba-bb13-0410-a3b6-899b003a5108

+++ trunk/themis/framework/ThemisTab.h
@@ -16,8 +16,8 @@
 class ThemisTab : public BTab
 {
 	public:
-									ThemisTab( BView *view, uint uid );
-		
+									ThemisTab( BView *view, int16 uid );
+									
 		virtual void				DrawTab(
 										ThemisTabView* owner,
 										BRect frame,
@@ -27,10 +27,10 @@ class ThemisTab : public BTab
 										ThemisTabView* owner,
 										BRect frame );
 										
-		uint						UniqueID();
+		int16						UniqueID();
 		
 	private:
-		uint						fUniqueID;
+		int16						fUniqueID;
 };
 
 #endif
